.github/skills/daily-market-brief/requirements.txt
PyYAML>=6.0
requests>=2.32.0
feedparser>=6.0.11
pytest>=8.3.0
tushare>=1.4.21
akshare>=1.16.0
jsonschema>=4.23.0